fixed logout
